[#87] remove resizeImage calls from index pages
